Move color.py inside layout/
