+++ pygyw/layout/__init__.py
@@ -3,3 +3,4 @@ from . import fonts
 from . import helpers
 from . import icons
 from . import settings
+from . import color
